+++ expression/expression.go
@@ -23,8 +23,19 @@ import (
 	"github.com/pingcap/tidb/context"
 	"github.com/pingcap/tidb/model"
 	"github.com/pingcap/tidb/mysql"
+	"github.com/pingcap/tidb/terror"
 	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/types"
+)
+
+// Error instances.
+var (
+	ErrInvalidOperation = terror.ClassExpression.New(CodeInvalidOperation, "invalid operation")
+)
+
+// Error codes.
+const (
+	CodeInvalidOperation terror.ErrCode = 1
 )
 
 // EvalAstExpr evaluates ast expression directly.
@@ -240,30 +251,6 @@ func EvaluateExprWithNull(ctx context.Context, schema Schema, expr Expression) (
 	default:
 		return x.Clone(), nil
 	}
-}
-
-// ResultFieldsToSchema converts slice of result fields to schema.
-func ResultFieldsToSchema(fields []*ast.ResultField) Schema {
-	schema := make(Schema, 0, len(fields))
-	for i, field := range fields {
-		colName := field.ColumnAsName
-		if colName.L == "" {
-			colName = field.Column.Name
-		}
-		tblName := field.TableAsName
-		if tblName.L == "" {
-			tblName = field.Table.Name
-		}
-		col := &Column{
-			ColName:  colName,
-			TblName:  tblName,
-			DBName:   field.DBName,
-			RetType:  &field.Column.FieldType,
-			Position: i,
-		}
-		schema = append(schema, col)
-	}
-	return schema
 }
 
 // TableInfo2Schema converts table info to schema.
